Improved code
